Refactored Gulp tasks to exist within a tasks folder
